chatbot-backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
apscheduler>=3.10.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
